feat: implement new design for brand profile
