+++ packages/component/src/category/CategoriesPage.tsx
@@ -97,7 +97,7 @@ const CategoriesPage = ({ categoryId, isMobileView }: CategoriesPageProps) => {
         onValueChange={(e) => setActiveTab(e as CATEGORY_PAGE_TABS)}
         className="flex h-full w-full flex-col bg-alpha-white"
       >
-        <TabsList className="sticky top-0 z-50 grid w-full grid-cols-4 bg-alpha-white font-medium sm:z-0  md:flex">
+        <TabsList className="sticky top-0 z-50 mb-5 grid w-full grid-cols-4 bg-alpha-white font-medium sm:z-0  md:flex">
           <TabsTrigger
             className="py-4"
             value={CATEGORY_PAGE_TABS.SUBCATEGORIES}
@@ -134,17 +134,17 @@ const CategoriesPage = ({ categoryId, isMobileView }: CategoriesPageProps) => {
             ""
           )}
         </TabsContent>
-        <TabsContent className="sm:pt-5" value={CATEGORY_PAGE_TABS.ORDERS}>
+        <TabsContent value={CATEGORY_PAGE_TABS.ORDERS}>
           <CategoriesPublicOrders
             publicPreOrders={publicPreOrders}
             isMobileView={isMobileView}
             categoryId={categoryId}
           />
         </TabsContent>
-        <TabsContent className="sm:pt-5" value={CATEGORY_PAGE_TABS.BRANDS}>
+        <TabsContent value={CATEGORY_PAGE_TABS.BRANDS}>
           <BrandsList args={brandsArgs} isMobileView={isMobileView} />
         </TabsContent>
-        <TabsContent className="sm:pt-5" value={CATEGORY_PAGE_TABS.PRODUCTS}>
+        <TabsContent value={CATEGORY_PAGE_TABS.PRODUCTS}>
           <ProductList
             isMobileView={isMobileView}
             args={productArgs}
