+++ packages/component/src/desktop/ListHeader.tsx
@@ -1,13 +1,19 @@
+import clsx from "clsx"
 import useTranslation from "next-translate/useTranslation"
 
 import ItemsCount from "../ItemsCount"
 
-type Props = { total?: number; listName: string }
+type Props = { borderBottom?: boolean; total?: number; listName: string }
 
-const ListHeader = ({ total, listName }: Props) => {
+const ListHeader = ({ borderBottom: borderBottom, total, listName }: Props) => {
   const { t } = useTranslation()
   return (
-    <div className="flex items-center justify-between border-b px md:px-0 md:pb-5">
+    <div
+      className={clsx(
+        "flex items-center justify-between  px md:px-0 md:pb-5",
+        borderBottom && "border-b"
+      )}
+    >
       <span className="text-lg font-medium">
         {t("common:entity_list", { entity: t(`common:${listName}`) })}
       </span>
